Implement thread-safe queue to be used in WriterThread
Looks like there's lock contention on stagedRequests
in WriterThread: flushing items to the socket,
receiving acknowledgements, and adding new items all
acquire the same mutex, which is silly.

ThreadSafeQueue supports poppers and pushers which
don't block each other, and efficient, almost-lock-free
scanning of the entire contents, using a linked list
of memory blocks.

To be used in WriterThread, and maybe a few other
places.