TEST: Add various scenarios of shutdowns and journal replays
We do mixed journal replay tests ie. in case of shutdown where we've not caught
up with all the responses, we restore from persistency. We test the various
conditions where we try to change the persistency layer before restarts, which
is what we expect to happen when we actually do this in production where a
serial -> parallel happens. But also test the inverse case in case we find
inconsistencies and need to switch to the earlier behaviour

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>